Refactor cache module and improve directory entry handling
- Changed `FileEntry::new` and `DirectoryEntry::new` methods from private to public.
- Renamed `SubdirectoryEntry` to `DirectoryEntry` for clarity.
- Added `get_dir_entry_serialized` method to `EntryType` for serialization of directory entries.
- Updated `DirectoryLookupContext` to use `DirectoryEntry` instead of `SubdirectoryEntry`.
- Removed unused `name_to_entry` HashMap and related logic from `DirectoryLookupContext`.
- Simplified event handling in `handle_fs_events` by batching invalidations and improving clarity.
- Introduced `execute_invalidations` function to handle cache invalidation logic.
- Changed `WINDOWS_MAX_PATH` type from `u64` to `u16` for proper representation.

+++ src/utils/windows.rs
@@ -20,5 +20,5 @@ pub mod time {
 }
 
 pub mod file {
-    pub const WINDOWS_MAX_PATH: u64 = 260;
+    pub const WINDOWS_MAX_PATH: u16 = 260;
 }
